refactor: update selector logic for isBrushing
