+++ packages/charts/src/chart_types/heatmap/state/selectors/get_cursor_pointer.ts
@@ -8,11 +8,11 @@
 
 import { CSSProperties } from 'react';
 
-import { isBrushingSelector } from './is_brushing';
 import { getPickedShapes, hasPicketVisibleCells } from './picked_shapes';
 import { DEFAULT_CSS_CURSOR } from '../../../../common/constants';
 import { createCustomCachedSelector } from '../../../../state/create_selector';
 import { getTooltipInteractionState } from '../../../../state/selectors/get_tooltip_interaction_state';
+import { isBrushingSelector } from '../../../../state/selectors/is_brushing';
 
 /** @internal */
 export const getPointerCursorSelector = createCustomCachedSelector(
