+++ packages/charts/src/chart_types/heatmap/state/selectors/get_highlighted_area.ts
@@ -8,8 +8,8 @@
 
 import { getHeatmapSpecSelector } from './get_heatmap_spec';
 import { getPerPanelHeatmapGeometries } from './get_per_panel_heatmap_geometries';
-import { isBrushingSelector } from './is_brushing';
 import { createCustomCachedSelector } from '../../../../state/create_selector';
+import { isBrushingSelector } from '../../../../state/selectors/is_brushing';
 
 /**
  * @internal
