+++ packages/charts/src/chart_types/xy_chart/renderer/dom/highlighter.tsx
@@ -14,6 +14,7 @@ import { GlobalChartState } from '../../../../state/chart_state';
 import { getChartRotationSelector } from '../../../../state/selectors/get_chart_rotation';
 import { getChartThemeSelector } from '../../../../state/selectors/get_chart_theme';
 import { InitStatus, getInternalIsInitializedSelector } from '../../../../state/selectors/get_internal_is_intialized';
+import { isBrushingSelector } from '../../../../state/selectors/is_brushing';
 import { getColorFromVariant, Rotation } from '../../../../utils/common';
 import { Dimensions } from '../../../../utils/dimensions';
 import { isPointGeometry, IndexedGeometry, PointGeometry } from '../../../../utils/geometry';
@@ -141,7 +142,7 @@ const mapStateToProps = (state: GlobalChartState): HighlighterProps => {
     initialized: true,
     chartId,
     zIndex,
-    isBrushing: state.interactions.pointer.dragging,
+    isBrushing: isBrushingSelector(state),
     highlightedGeometries: getHighlightedGeomsSelector(state),
     chartTransform: computeChartTransformSelector(state),
     chartDimensions: computeChartDimensionsSelector(state).chartDimensions,
